Updated module and playbook
